% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCDF.R
\name{CCDF}
\alias{CCDF}
\title{Cumulative Complementary Density Function (CCDF)}
\usage{
CCDF(x, include.min = FALSE)
}
\arguments{
\item{x}{numeric vector to calculate CCDF of}

\item{include.min}{whether to include a first point z=min(x)-1, which will have P(x>z)=1}
}
\value{
3 row matrix with a first row with the values of the vector v, the second one with
P(x>v) and the third one with the absolute count N of x>v.
}
\description{
This function computes the CCDF of the values in x, which is P(x>v),
i.e. the probability of finding a value strictly larger than v in the input vector x.
}
\examples{
xs <- floor(runif(min=0,max=10,n=100))
CCDF(xs)

}
\keyword{density,}
\keyword{power-law}

