% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCCDF.R
\name{plotCCDF}
\alias{plotCCDF}
\title{Plot a Cumulative Complementary Density Function (CCDF)}
\usage{
plotCCDF(x, include.min = FALSE, counts = FALSE, log = "xy", type = "l",
  xlab = "x", ylab = "P(X>x)", cex.lab = 1.3, cex.axis = 1.3, lwd = 2,
  ...)
}
\arguments{
\item{x}{numeric vector to calculate CCDF of}

\item{include.min}{whether to include a first point z=min(x)-1, which will have P(x>z)=1 (default FALSE)}

\item{counts}{whether to plot absolute counts instead of probability (default FALSE)}

\item{log}{character string determining the axes to display logarithmically (default "xy", both axes)}

\item{type}{of plot (default line "l")}

\item{xlab}{label of x axis (default "x")}

\item{ylab}{label of y axis (default "P(X>x)")}

\item{cex.lab}{size of the axis label (default 1.3)}

\item{cex.axis}{size of the axis values (default 1.3)}

\item{lwd}{width of the plot line (default 2)}

\item{...}{additional parameters for plot function}
}
\description{
This function calculates and plots a CCDF with logarithmic axes by default
}
\examples{
xs <- floor(runif(min=0,max=10,n=100))
plotCCDF(xs)

}
\keyword{density,}
\keyword{power-law}

